requirements.txt
google_cloud_aiplatform==1.67.1
pandas==2.2.3
protobuf==3.20.3
streamlit==1.32.0
streamlit==1.38.0
~treamlit==1.37.1
